cmd,lint: move lint/flagparser to cmd/internal/flagparser
